# [v2] Require dt ids for errors if one is given.