Change json spec to require `transaction_id`, `trace_id` and `parent_id` as soon as one of them is given.

Enforcing this conditional requirement on JSON spec level, decreases performance:
```
benchmark                                                   old ns/op     new ns/op     delta
BenchmarkStreamProcessor/errors.ndjson-4                    147158        163343        +11.00%

benchmark                                                   old MB/s     new MB/s     speedup
BenchmarkStreamProcessor/errors.ndjson-4                    29.83        26.88        0.90x

benchmark                                                   old allocs     new allocs     delta
BenchmarkStreamProcessor/errors.ndjson-4                    833            909            +9.12%

benchmark                                                   old bytes     new bytes     delta
BenchmarkStreamProcessor/errors.ndjson-4                    45311         47782         +5.45%
```

The alternative is to move the check to the go code in the `model/error/event.go`. This would mean basically no overhead. However, we decided to use JSON schema and allow agent devs to run their tests against it, so I didn't simply want to move the check out of the schema.