@felixbarny 2018-09-19
I'm ok with both