@roncohen 2018-09-19
I'm also OK with both. The following should work as well, but it didn't change the benchmark much: 

```diff
                     "maxLength": 1024
                 }
             },
+            "required": ["id"],
             "oneOf": [
-                { "required": ["id", "trace_id", "transaction_id", "parent_id"] },
+                { "required": ["trace_id", "transaction_id", "parent_id"] },
                 { "allOf": [
-                    { "required": ["id"] },
                     { "not": { "required": ["trace_id"] }},
                     { "not": { "required": ["transaction_id"] }},
                     { "not": { "required": ["parent_id"] }}
```