@simitt 2018-09-19
@roncohen there are a view different approaches on how to build the requirements, but all cases I checked showed roughly the same performance. Thanks for trying it out though.