@roncohen 2018-09-19
@simitt cool. Is there a reason the "id" separate like in the example I showed? From my perspective it seems simpler to understand