@Qard 2018-09-19
Do the `{ "not": { "required": ... } }` bits make it _optional_ or required that it _not be present_?