@simitt 2018-09-20
After digging into this some more, I updated to use `if/then` statements, for better readability and they also perform a lot better: 
```
benchmark                                                   old ns/op     new ns/op     delta
BenchmarkStreamProcessor/errors.ndjson-4                    147429        155261        +5.31%

benchmark                                                   old MB/s     new MB/s     speedup
BenchmarkStreamProcessor/errors.ndjson-4                    29.78        28.27        0.95x

benchmark                                                   old allocs     new allocs     delta
BenchmarkStreamProcessor/errors.ndjson-4                    833            869            +4.32%

benchmark                                                   old bytes     new bytes     delta
BenchmarkStreamProcessor/errors.ndjson-4                    45307         46520         +2.68%
`
``